Simple benchmark: restructure files
